requirements.txt
accelerate==1.7.0
aiofiles==24.1.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
av==14.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
defusedxml==0.7.1
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.3.2
gradio==5.29.0
gradio_client==1.10.0
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.31.2
hydra-core==1.3.2
idna==3.10
iopath==0.1.10
Jinja2==3.1.6
kiwisolver==1.4.8
lang-sam @ git+https://github.com/luca-medeiros/lang-segment-anything.git@e9af744d999d85eb4d0bd59a83342ecdc2bd2461
litserve==0.2.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
nvidia-cublas-cu12==12.4.2.65
nvidia-cuda-cupti-cu12==12.4.99
nvidia-cuda-nvrtc-cu12==12.4.99
nvidia-cuda-runtime-cu12==12.4.99
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.0.44
nvidia-curand-cu12==10.3.5.119
nvidia-cusolver-cu12==11.6.0.99
nvidia-cusparse-cu12==12.3.0.142
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.4.99
omegaconf==2.3.0
opencv-python-headless==4.10.0.84
orjson==3.10.18
packaging==25.0
pandas==2.2.3
pillow==11.2.1
portalocker==3.1.1
psutil==7.0.0
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
qwen-vl-utils==0.0.11
regex==2024.11.6
requests==2.32.3
rich==14.0.0
ruff==0.11.10
safehttpx==0.1.6
safetensors==0.5.3
SAM-2 @ git+https://github.com/facebookresearch/segment-anything-2@c2ec8e14a185632b0a5d8b161928ceb50197eddc
scipy==1.15.3
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
supervision==0.23.0
sympy==1.14.0
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.4.1+cu124
torchvision==0.19.1+cu124
tqdm==4.67.1
transformers==4.51.3
triton==3.0.0
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.20.0
watchfiles==1.0.5
websockets==15.0.1
